split AutoCompleteField.spec into two files
